+++ src/classes/numberGenerator.js
